refactor : directory structure for better organization
