--- app/src/main/java/online/partyrun/partyrunapplication/common/Constants.kt
@@ -1,8 +1,0 @@
-package online.partyrun.partyrunapplication.common
-
-import online.partyrun.partyrunapplication.BuildConfig
-
-object Constants {
-    const val BASE_URL = BuildConfig.TEST_BASE_URL
-
-}
