+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/main/Test2Screen.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.presentation.main.nav_test_screen
+package online.partyrun.partyrunapplication.presentation.main
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.Button
