+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/main/Test3Screen.kt
@@ -1,0 +1,25 @@
+package online.partyrun.partyrunapplication.presentation.main
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun Test3Screen(
+    onSignOut: () -> Unit
+) {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Button(onClick = onSignOut) {
+            Text(text = "Sign out")
+        }
+    }
+}
