--- app/src/main/java/online/partyrun/partyrunapplication/presentation/main/nav_test_screen/Test3Screen.kt
@@ -1,8 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.main.nav_test_screen
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Test3Screen() {
-
-}
