+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/test/TestViewModel.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.presentation.screens.test
+package online.partyrun.partyrunapplication.presentation.test
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
