+++ app/src/main/java/online/partyrun/partyrunapplication/utils/Constants.kt
@@ -1,0 +1,9 @@
+package online.partyrun.partyrunapplication.utils
+
+import online.partyrun.partyrunapplication.BuildConfig
+
+object Constants {
+    const val TEST_BASE_URL = BuildConfig.TEST_BASE_URL
+    const val FB_GOOGLE_WEB_CLIENT_ID = BuildConfig.FB_GOOGLE_WEB_CLIENT_ID
+
+}
